Change kinds to C_<kind>

+++ sources/ncf90api_readgrid.f90
@@ -36,7 +36,7 @@ subroutine readgrid2d_byte(ifile, idata)
   character(*) :: ifile
   type(nc2d_byte) :: idata
 
-  integer(kind=4) :: ncid, varid, xvarid, yvarid, vartype, nofill
+  integer(kind=C_INT) :: ncid, varid, xvarid, yvarid, vartype, nofill
 
   call ncoords(ifile, idata)
 
@@ -64,7 +64,7 @@ subroutine readgrid2d_short(ifile, idata)
   character(*) :: ifile
   type(nc2d_short) :: idata
 
-  integer(kind=4) :: ncid, varid, xvarid, yvarid, vartype, nofill
+  integer(kind=C_INT) :: ncid, varid, xvarid, yvarid, vartype, nofill
 
   call ncoords(ifile, idata)
 
@@ -92,7 +92,7 @@ subroutine readgrid2d_int(ifile, idata)
   character(*) :: ifile
   type(nc2d_int) :: idata
 
-  integer(kind=4) :: ncid, varid, xvarid, yvarid, vartype, nofill
+  integer(kind=C_INT) :: ncid, varid, xvarid, yvarid, vartype, nofill
 
   call ncoords(ifile, idata)
 
@@ -120,7 +120,7 @@ subroutine readgrid2d_float(ifile, idata)
   character(*) :: ifile
   type(nc2d_float) :: idata
 
-  integer(kind=4) :: ncid, varid, xvarid, yvarid, vartype, nofill
+  integer(kind=C_INT) :: ncid, varid, xvarid, yvarid, vartype, nofill
 
   call ncoords(ifile, idata)
 
@@ -148,7 +148,7 @@ subroutine readgrid2d_double(ifile, idata)
   character(*) :: ifile
   type(nc2d_double) :: idata
 
-  integer(kind=4) :: ncid, varid, xvarid, yvarid, vartype, nofill
+  integer(kind=C_INT) :: ncid, varid, xvarid, yvarid, vartype, nofill
 
   call ncoords(ifile, idata)
